split in some component and create AuthFlow component
